+++ application/src/main/resources/ecc.yml
@@ -35,77 +35,76 @@ connection:
     ## Configuration to define Agent Strategy and hosts for ecChronos
     ## to connect to. The application will use the configurations
     ## specified below, connecting to the listed hosts;
-    agent:
-      ## Each ecChronos instance must have unique name.
-      ## (instanceName: unique identifier), that will be used
-      ## as ecchronos_id (partition key in nodes_sync table).
-      instanceName: unique_identifier
-      ## Define the Agent strategy, it can be
-      ## - datacenterAware;
-      ## - rackAware; and
-      ## - hostAware.
-      type: datacenterAware
-      ## Specifies the datacenter that is considered "local" by the load balancing policy,
-      ## The specified datacenter should match with the contact point datacenter
-      localDatacenter: datacenter1
-      ## Initial contact points list for ecChronos
-      ## to establish first connection with Cassandra.
-      contactPoints:
+    ## Each ecChronos instance must have unique name.
+    ## (instanceName: unique identifier), that will be used
+    ## as ecchronos_id (partition key in nodes_sync table).
+    instanceName: unique_identifier
+    ## Define the Agent strategy, it can be
+    ## - datacenterAware;
+    ## - rackAware; and
+    ## - hostAware.
+    type: datacenterAware
+    ## Specifies the datacenter that is considered "local" by the load balancing policy,
+    ## The specified datacenter should match with the contact point datacenter
+    localDatacenter: datacenter1
+    ## Initial contact points list for ecChronos
+    ## to establish first connection with Cassandra.
+    contactPoints:
+      - host: 127.0.0.1
+        port: 9042
+      - host: 127.0.0.2
+        port: 9042
+
+    ## Configuration to define datacenters for ecChronos
+    ## to connect to, datacenterAware enable means that
+    ## ecChronos will be responsible for all nodes in the
+    ## datacenter list.
+    datacenterAware:
+      datacenters:
+        - name: datacenter1
+        - name: datacenter2
+    ## Configuration to define racks for ecChronos
+    ## to connect to, rackAware enable means that
+    ## ecChronos will be responsible for all nodes in the
+    ## rack list, this configuration is designed to manage
+    ## racks on the same datacenter.
+    rackAware:
+      racks:
+        - datacenterName: datacenter1
+          rackName: rack1
+        - datacenterName: datacenter1
+          rackName: rack2
+    ## Configuration to define hosts for ecChronos
+    ## to connect to, hostAware enable means that
+    ## ecChronos will be responsible just for the
+    ## specified hosts list.
+    hostAware:
+      hosts:
         - host: 127.0.0.1
           port: 9042
         - host: 127.0.0.2
           port: 9042
-
-      ## Configuration to define datacenters for ecChronos
-      ## to connect to, datacenterAware enable means that
-      ## ecChronos will be responsible for all nodes in the
-      ## datacenter list.
-      datacenterAware:
-        datacenters:
-          - name: datacenter1
-          - name: datacenter2
-      ## Configuration to define racks for ecChronos
-      ## to connect to, rackAware enable means that
-      ## ecChronos will be responsible for all nodes in the
-      ## rack list, this configuration is designed to manage
-      ## racks on the same datacenter.
-      rackAware:
-        racks:
-          - datacenterName: datacenter1
-            rackName: rack1
-          - datacenterName: datacenter1
-            rackName: rack2
-        ## Configuration to define hosts for ecChronos
-      ## to connect to, hostAware enable means that
-      ## ecChronos will be responsible just for the
-      ## specified hosts list.
-      hostAware:
-        hosts:
-          - host: 127.0.0.1
-            port: 9042
-          - host: 127.0.0.2
-            port: 9042
-          - host: 127.0.0.3
-            port: 9042
-          - host: 127.0.0.4
-            port: 9042
-      # Schedule configuration for the ReloadSchedulerService, responsible for monitoring
-      # Cassandra topology changes during execution. If a topology change is detected,
-      # the internal topology state is automatically refreshed.
-      reloadSchedule:
-        ## Initial delay before the RetrySchedulerService starts after the application launches.
-        ## By default, this service would trigger after 1 day and represents the wait time before the service first runs
-        initialDelay: 1
-        ## Fixed delay between subsequent executions of the RetrySchedulerService.
-        ## This value represents the delay time (by default 1 day) between when the service completes one execution
-        ## and the time it will be triggered again for another execution.
-        ## During each execution, the service will check for unavailable nodes and attempt reconnections as per the retry logic.
-        fixedDelay: 1
-        ## unit can be milliseconds, seconds, minutes, hours, days.
-        unit: days
-      # Retry policy used during the creation of a CqlSession.
-      # This policy defines how connection attempts should be retried in case
-      # of temporary failures when establishing a session with Cassandra.
+        - host: 127.0.0.3
+          port: 9042
+        - host: 127.0.0.4
+          port: 9042
+    # Schedule configuration for the ReloadSchedulerService, responsible for monitoring
+    # Cassandra topology changes during execution. If a topology change is detected,
+    # the internal topology state is automatically refreshed.
+    reloadSchedule:
+      ## Initial delay before the RetrySchedulerService starts after the application launches.
+      ## By default, this service would trigger after 1 day and represents the wait time before the service first runs
+      initialDelay: 1
+      ## Fixed delay between subsequent executions of the RetrySchedulerService.
+      ## This value represents the delay time (by default 1 day) between when the service completes one execution
+      ## and the time it will be triggered again for another execution.
+      ## During each execution, the service will check for unavailable nodes and attempt reconnections as per the retry logic.
+      fixedDelay: 1
+      ## unit can be milliseconds, seconds, minutes, hours, days.
+      unit: days
+    # Retry policy used during the creation of a CqlSession.
+    # This policy defines how connection attempts should be retried in case
+    # of temporary failures when establishing a session with Cassandra.
     retryPolicy:
       ## Max number of attempts ecChronos will try to connect with Cassandra.
       maxAttempts: 5
